+++ packages/react-core/src/provider/index.ts
@@ -1,0 +1,1 @@
+export {useFlowQueryClient, FlowQueryClientContext} from "./FlowQueryClient"