+++ packages/react-core/src/utils/index.ts
@@ -1,0 +1,2 @@
+export * from "./deepEqual"
+export * from "./flowscan"