+++ packages/react-sdk/src/components/TransactionButton.tsx
@@ -1,5 +1,5 @@
 import React from "react"
-import {useFlowMutate} from "../hooks"
+import {useFlowMutate} from "@onflow/react-core"
 import {Button, ButtonProps} from "./internal/Button"
 import {StyleWrapper} from "./internal/StyleWrapper"
 import {UseMutationOptions} from "@tanstack/react-query"